+++ tests/test_version.py
@@ -1,7 +1,9 @@
 import cascade_fer
 import logging
+
 logging.basicConfig(level=logging.INFO)
+
 
 def test_version():
     logging.info(f"Version: {cascade_fer.__version__}")
-    assert cascade_fer.__version__ == "0.1.1"
+    assert cascade_fer.__version__ == "0.1.2"