+++ Changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 All major changes are listed here
+# [0.1.2] Nov 7th 2021
+# Added
+- [] A stand-alone fer that takes a raw input image and outputs the dominant emotion over a duration
+
+# Changed
+- [] Simplify live facial emotion detection using the stand-alone fer
+
 # [0.1.1] Oct 24th 2021
 ## Added
 - [x] Live facial emotion detection